creates: new app restapi, redirects host/rest/ to this app, adds this in settings 	1. moves to_json_style(), read_regex(), read_family(), toggle_is_achieved() 		from webinterface.views -> restapi.views 		and similarly their urls
