+++ escapeerrands/urls.py
@@ -6,6 +6,10 @@ urlpatterns = [
     # host/admin/
     url(r'^admin/', admin.site.urls),
 
+    # REST Api
+    # host/rest/
+    url(r'^rest/', include('restapi.urls')),
+
     # Web Interface
     # host/
     url(r'^', include('webinterface.urls')),
