+++ webinterface/templates/webinterface/base.html
@@ -4,13 +4,12 @@
     <meta charset="UTF-8">
     {% load static %}
     {# JQuery #}
-    <script type="text/javascript" src="{% static 'webinterface/jquery.min.js' %}"></script>
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
     {# Materialize #}
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.100.1/css/materialize.min.css">
 {#    <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.100.1/js/materialize.min.js"></script>#}
-    {# Bucketsjs #}
-{#    <script type="text/javascript" src="{% static 'webinterface/buckets.min.js' %}"></script>#}
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     {# Extra Headers #}
     {% block extra_headers %}{% endblock %}
     {# Icon and Title #}
