+++ webinterface/templates/webinterface/goal/scratch.html
@@ -1,0 +1,31 @@
+{% extends 'webinterface/base.html' %}
+{% block title %}Goals at a Glance{% endblock %}
+{% load static %}
+{% block extra_headers %}
+    <script src="http://d3js.org/d3.v3.min.js"></script>
+    <script src="{% static 'webinterface/jsnetworkx.js' %}"></script>
+{% endblock %}
+{% block body %}
+    <div class="container-fluid">
+        <div id="canvas" style="height: 80vh"></div>
+        <script>
+            $(document).ready(function () {
+                var G = new jsnx.Graph();
+                G.addNode(1, {label: 'foo'});
+                G.addNode(2, {label: 'bar'});
+                G.addNode(3, {label: 'baz'});
+                G.addEdgesFrom([[1, 2], [2, 3]]);
+
+                jsnx.draw(G, {
+                        'element': '#canvas',
+                        'node_style': {
+                            'stroke': '#333',
+                            'fill': '#999000',
+                            'cursor': 'pointer'
+                        }
+                    },
+                    true);
+            });
+        </script>
+    </div>
+{% endblock %}
